refactor(edit): identity grid layout + responsive mobile spans (OSC-125)
- Portrait spans 2 rows top-left on wide; small square (span 3) on mobile
- Alignment/vitals reflow gap-free; comboboxes go full-width (span 12) on mobile
- Keep the 12-col grid at narrow widths, override wide spans instead of collapsing to 2-up
- Extract HitDiceField into its own module; PortraitField forwards a style prop
